[DDW-926]: Update filter dialog position handling on transactions screen
