+++ source/renderer/app/components/wallet/transactions/FilterDialog.scss
@@ -197,7 +197,7 @@
 }
 
 :global {
-  .small-height-screen {
+  .small-height-for-date-picker {
     .FilterDialog_component {
       .rdtPicker {
         left: -281px;
